+++ src/main/java/Codigo/Main.java
@@ -16,7 +16,7 @@ public class Main {
         
         //System.out.println(cmol.S_cmol((float)0.15,(float)5.76,(float)1.63));
         //System.out.println(cmol.CTC_cmol((float)0.15,(float)5.76,(float)1.63, (float)5.35));
-        System.out.println(cmol.V_atual((float)0.15,(float)5.76,(float)1.63, (float)5.35));
+        System.out.println(cmol.V_atual(0.15,5.76,1.63, 5.35));
         
         //--------------------------------------------
